Update README.md with first project's info
This repository is expected to hold, as of now, all of the practice
projects I am assigned via The Odin Project's curriculum.

Established a basic framework for describing these projects individually
in README.md. TOP recommends reflecting on Project #1 both before
working on & after completing it, and I plan to repeat this for any
future projects as well. Hopefully the template helps!

These early commits will be fairly verbose as I learn how to write them.

If you're reading this: I apologize!